[140722] 통신 구조 변경
- 통신 구조(Serial) 변경(1.0.0.5로 버전 변경)

git-svn-id: svn://192.168.30.90:4000/%5BSRC%5D/%5BSVN%5D%20ECS-RD12-16%20-%20WLF%20Monitor%20-%20STM32_APP@47 8c143bec-4d7d-2e4e-a212-1cf3cf7e8ea1

+++ Platform/TaeHa - ECS/INC/WL9F_Monitor_Variable.h
@@ -47,7 +47,7 @@ USARTx_INDEX 				WL9FM_USART_INDEX;
 USARTx_FILE_RX_DATA			WL9FM_USART_RX_FILE_DATA;
 USARTx_FILE_RX_DATA			WL9FM_USART_FILE_DATA;
 
-uint8_t                    	KeySwitch_Value;            //  GPIO Key Value
+uint32_t                    	KeySwitch_Value;            //  GPIO Key Value
 uint32_t 					LAMP_Update_Data;			//	LAMP Value
 
 /* Private function prototypes -----------------------------------------------*/
